+++ src/onegov.town/onegov/town/models/form.py
@@ -1,9 +1,10 @@
 from onegov.form.models import FormDefinition
+from onegov.town.models.mixins import HiddenMetaMixin
 
 
-class BuiltinFormDefinition(FormDefinition):
+class BuiltinFormDefinition(FormDefinition, HiddenMetaMixin):
     __mapper_args__ = {'polymorphic_identity': 'builtin'}
 
 
-class CustomFormDefinition(FormDefinition):
+class CustomFormDefinition(FormDefinition, HiddenMetaMixin):
     __mapper_args__ = {'polymorphic_identity': 'custom'}